Second split done, stairway by itself in both versions now
