+++ config/splat.us.stno3.yaml
@@ -134,7 +134,7 @@ segments:
       - [0x26638, .data, tile_data] # tile definitions
       - [0x2EA68, .data, gen/sprites]
       - [0x373E0, .rodata, e_castle_door]
-      - [0x37400, .rodata, entrance_stage_entities_2] # EntityStairwayPiece
+      - [0x37400, .rodata, e_stairway]
       - [0x37414, .rodata, 3C4EC]
       - [0x3742C, .rodata, 3E134] #EntityFlyingOwlAndLeaves
       - [0x37440, .rodata, cutscene] # actor names
@@ -170,7 +170,8 @@ segments:
       - [0x394F0, c, e_cavern_door]
       - [0x39C44, c, entrance_weights]
       - [0x3A348, c, entrance_stage_entities_3]
-      - [0x3AEF0, c, entrance_stage_entities_2]
+      - [0x3AEF0, c, e_stairway]
+      - [0x3B548, c, entrance_stage_entities_2]
       - [0x3BF38, c, e_heartroom]
       - [0x3C3FC, c, e_tilemap_shuffler]
       - [0x3C4EC, c]
