+++ config/splat.us.stnp3.yaml
@@ -137,7 +137,7 @@ segments:
       - [0x21CF4, .data, tile_data] # tile definitions
       - [0x2A124, .data, gen/sprites]
       - [0x31EA0, .rodata, e_castle_door]
-      - [0x31EB0, .rodata, entrance_stage_entities_2] # EntityStairwayPiece
+      - [0x31EB0, .rodata, e_stairway]
       - [0x31EBC, .rodata, e_explosion_puff_opaque] # EntityExplosionPuffOpaque
       - [0x31ED0, .rodata, slogra] # EntitySlogra .rodata, 36990
       - [0x31EDC, .rodata, gaibon] # EntityGaibon .rodata, 36990
@@ -171,7 +171,8 @@ segments:
       - [0x33D24, c, e_cavern_door]
       - [0x344B4, c, entrance_weights]
       - [0x34BE8, c, entrance_stage_entities_3]
-      - [0x35790, c, entrance_stage_entities_2]
+      - [0x35790, c, e_stairway]
+      - [0x35DE8, c, entrance_stage_entities_2]
       - [0x35F60, c, e_heartroom]
       - [0x3644C, c, e_tilemap_shuffler]
       - [0x3653C, c, e_explosion_puff_opaque]
