+++ src/st/entrance_stage_entities_2.h
@@ -1,188 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-// Stairway piece you can break before Death encounter
-void EntityStairwayPiece(Entity* self, u8 arg1, u8 arg2, u8 arg3) {
-    Primitive *prim, *prim2, *prim3;
-    Entity* newEntity;
-    Collider collider;
-    s16 primIndex;
-    s32 temp;
-    s16 x, y;
-    u8 v1;
-    s32 i;
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(g_EInitInteractable);
-        self->hitboxWidth = 8;
-        self->hitboxHeight = 8;
-        self->posX.i.hi = 1432 - g_Tilemap.scrollX.i.hi;
-        self->posY.i.hi = 200 - g_Tilemap.scrollY.i.hi;
-        self->hitPoints = 16;
-        if (g_CastleFlags[DEATH_STAIRWAY_BROKEN]) {
-            self->hitboxState = 0;
-            g_Tilemap.fg[0x4D9] = 0x3EE;
-            g_Tilemap.fg[0x539] = 0x3D2;
-            self->step = 32;
-            break;
-        }
-        self->hitboxState = 2;
-        g_Tilemap.fg[0x4D9] = 0x408;
-        g_Tilemap.fg[0x539] = 0x40D;
-        break;
-
-    case 1:
-        if (self->hitFlags != 0) {
-            g_api.PlaySfx(SFX_STOMP_SOFT_A);
-        }
-
-        if (self->flags & FLAG_DEAD) {
-            self->step++;
-        }
-        break;
-
-    case 2:
-        g_api.PlaySfx(SFX_WALL_DEBRIS_B);
-        g_Tilemap.fg[0x4D9] = 0x3EE;
-        g_Tilemap.fg[0x539] = 0x3D2;
-        g_CastleFlags[DEATH_STAIRWAY_BROKEN] = true;
-
-        newEntity = AllocEntity(&g_Entities[160], &g_Entities[192]);
-        if (newEntity != NULL) {
-            CreateEntityFromEntity(E_EQUIP_ITEM_DROP, self, newEntity);
-            newEntity->params = ITEM_TURKEY;
-        }
-
-        newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
-        if (newEntity != NULL) {
-            CreateEntityFromEntity(E_INTENSE_EXPLOSION, self, newEntity);
-            newEntity->params = 0x10;
-            newEntity->zPriority = self->zPriority + 1;
-            newEntity->posX.i.hi += 8;
-            newEntity->posY.i.hi += 8;
-        }
-
-        primIndex = g_api.AllocPrimitives(PRIM_GT4, 16);
-        if (primIndex == -1) {
-            DestroyEntity(self);
-            return;
-        }
-        prim = &g_PrimBuf[primIndex];
-        self->primIndex = primIndex;
-        self->ext.prim = prim;
-        self->flags |= FLAG_HAS_PRIMS;
-        UnkPolyFunc2(prim);
-        v1 = g_Tilemap.tileDef->gfxIndex[0x409];
-        arg1 = v1;
-        temp = g_Tilemap.tileDef->gfxPage[0x409];
-        prim->clut = g_Tilemap.tileDef->clut[0x409];
-        prim->tpage = temp + 8;
-        arg1 *= 16;
-        arg3 = 0xF;
-        arg3 = arg1 | arg3;
-        prim->u0 = prim->u2 = arg1;
-        arg2 = v1 & 0xF0 | 0xF;
-        prim->v0 = prim->v1 = v1 & 0xF0;
-        prim->u1 = prim->u3 = arg3;
-        prim->v2 = prim->v3 = arg2;
-        prim->next->x1 = self->posX.i.hi;
-        prim->next->y0 = self->posY.i.hi;
-        LOW(prim->next->u0) = 0xFFFF0000;
-        LOW(prim->next->r1) = 0xFFFF0000;
-        LOH(prim->next->r2) = 16;
-        LOH(prim->next->b2) = 16;
-        prim->priority = self->zPriority;
-        prim->drawMode = DRAW_UNK02;
-        self->step++;
-
-    case 3:
-        prim = self->ext.prim;
-        prim2 = prim->next;
-        prim2->tpage -= 0x20;
-        prim2 = prim->next;
-        LOW(prim2->r1) += 0x2000;
-        UnkPrimHelper(prim);
-        prim3 = prim->next;
-        x = prim3->x1;
-        y = prim3->y0;
-        g_api.CheckCollision(x, (s16)(y + 8), &collider, 0);
-        if (collider.effects & EFFECT_SOLID) {
-            self->posX.i.hi = x;
-            self->posY.i.hi = y - 4;
-            self->step++;
-        }
-        break;
-
-    case 4:
-        g_api.PlaySfx(SFX_WALL_DEBRIS_B);
-        newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
-        if (newEntity != NULL) {
-            CreateEntityFromEntity(E_EXPLOSION, self, newEntity);
-            newEntity->params = 0x11;
-            newEntity->zPriority = self->zPriority + 1;
-        }
-
-        for (i = 0; i < 6; i++) {
-            newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
-            if (newEntity != NULL) {
-                CreateEntityFromEntity(E_FALLING_ROCK, self, newEntity);
-                newEntity->params = Random() & 3;
-                if (newEntity->params == 3) {
-                    newEntity->params = 0;
-                }
-            }
-        }
-        DestroyEntity(self);
-    }
-}
-
-// falling rock with puff of smoke when it disappears. I think part of the
-// merman room breakable rock
-void EntityFallingRock(Entity* self) {
-    s32 animFrame = self->params & 0xF;
-    Collider collider;
-    Entity* newEntity;
-    s16 rndAngle;
-    s32 rnd;
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(g_EInitStInteractable);
-        self->animCurFrame = animFrame + 31;
-        self->rotY = 0x60;
-        self->rotX = 0x60;
-        self->drawFlags |= FLAG_DRAW_ROTZ | FLAG_DRAW_ROTY | FLAG_DRAW_ROTX;
-        rnd = (Random() & 0x1F) + 16;
-        rndAngle = (Random() * 6) + 0x900;
-        self->velocityX = rnd * rcos(rndAngle);
-        self->velocityY = rnd * rsin(rndAngle);
-        if (self->velocityX > 0) {
-            self->facingLeft = 1;
-        }
-        break;
-
-    case 1:
-        MoveEntity();
-        self->velocityY += FIX(0.125);
-        self->rotZ -= 0x20;
-
-        g_api.CheckCollision(
-            self->posX.i.hi, self->posY.i.hi + 8, &collider, 0);
-        if (collider.effects & EFFECT_SOLID) {
-            newEntity = AllocEntity(&g_Entities[224], &g_Entities[256]);
-            if (newEntity != NULL) {
-                CreateEntityFromEntity(E_INTENSE_EXPLOSION, self, newEntity);
-                newEntity->params = 0x10;
-                if (animFrame == 0) {
-                    newEntity->params = 0x13;
-                }
-            }
-            DestroyEntity(self);
-        }
-        break;
-    }
-}
-
 #if defined(STAGE_IS_NO3)
 static SVECTOR vec_negneg = {-0x100, 0, -0x100};
 static SVECTOR vec_posneg = {0x100, 0, -0x100};
